Added some userinfo and changed code a bit overall
