+++ Clients/GUI/Pages/Index.razor
@@ -1,6 +1,6 @@
 ﻿@page "/"
 
-<h3>Index</h3>
+<h3>Welcome</h3>
 
 @code {
