+++ Clients/GUI/Shared/NavMenu.razor
@@ -1,4 +1,4 @@
-﻿@inject Auth.AuthStateProvider AuthenticationProvider
+﻿@inject Authentication.AuthStateProvider AuthenticationProvider
 @inject NavigationManager NavigationManager
 
 <div class="top-row pl-4 navbar navbar-dark">
