+++ Clients/GUI/ViewModels/SignInViewModel.cs
@@ -1,11 +1,12 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace GUI.GUIModels
+namespace GUI.ViewModels
 {
     // Model for user input in SignIn page
-    public class SignInData
+    public class SignInViewModel
     {
         [Required (ErrorMessage = "Email is required")]
+        [EmailAddress]
         public string Email { get; set; }
 
         [Required (ErrorMessage = "Password is required")]
